Updated openstack/openstack Project: openstack/tripleo-incubator  270ccd1e5ec801afdb2f99a6a4470e65dc5f6d29
Do not set deploy_ramdisk and deploy_kernel as flavor keys

Setting deploy_ramdisk and deploy_kernel as flavor extra_specs is
deprecated from more than a cycle and not supported anymore. The deploy
kernel and ramdisk name are instead passed to register_nodes so that
they are set on the node itself in Ironic.

Co-authored-By: James Slagle <jslagle@redhat.com>
Co-authored-By: Derek Higgins <derekh@redhat.com>

Change-Id: I494e3579c3daad7bf13a8127cf83cc61f82f913a
Related-To: I0727126770e6c79c898560a34e91a10686696d82
Related-Bug: 1453076
Depends-On: Ie38d3186b45fe00e0577f4b0f8b6a971edb9827d