Updated openstack/openstack Project: openstack/oslo.messaging  ecb7803d5fc8c612ef12abd6aadc070596393882
rabbit: remove publisher classes

Each times we send a message with the rabbit driver we create useless
object that's not more that the kombu exchange and a wrapper method around
the kombu producer.

So, this change just creates the exchange not our useless custom Publisher
and move the wrapped methods into the Connection object.

Change-Id: Id221f4363d897cd904f7aeccbc90cbd288db2db1