Updated openstack/openstack Project: openstack/heat  a65d7e28c257c72cf179d18c2a3adbdf918872da
Rename OS::Heat::StructuredDeployments

Rename the StructuredDeployments resource to StructuredDeploymentGroup
to avoid awkward situations when talking about the plural form of the
StructuredDeployment resource. This has been agreed at the Vancouver
summit.

The existing 'StructuredDeployments' resource is kept for backwards
compatibility, but with a deprecated status and corresponding
deprecation message.

Closes-Bug: #1458008

Change-Id: I5d274acf17f222c493d10ded50f87e05a075b424