Updated openstack/openstack Project: openstack/cinder  b80cf733c4056b890a8da7629e9cb0f2162a5366
Add missing '-o' CLI option to VNX Cinder Driver

In some NaviSecCLI command, if -o is not given, the command may prompt
some message and ask for 'y' or 'n' for confirmation in some special
scenario, which is not working for Cinder.

This patch is to add the missing '-o' option to avoid these failures.

Change-Id: Ia665f6a20703218065aa4cf65b766def3daee290
Closes-Bug: #1453654