Updated openstack/openstack Project: openstack/horizon  46ca8a7095d8ed65d81381440705750765a81a96
ngReorg - Create dashboard-app

This is the last patch of this ngReorg series!!

This commit attempts to refactor the "hz" module and horizon.js
into "dashboard-app" because the code is actually bootstrapping
the default dashboard application, created using framework and
openstack_dashboard pieces. Ideally, this moves into a directory
that is a peer to openstack_dashboard, but for now, it is at
least grouped together with a name that reflects its purpose.

This was the last step in a larger effort to restructure the Angular
source. See https://review.openstack.org/#/c/176152/ for the
full set of planned changes.

The next refactoring effort (not addressed by this patch series)
is to restructure the Django templates so that
horizon.dashboard-app can be moved OUT of horizon/horizon.

Change-Id: If2cbc243ebfd0b2f1f834056e34b8e2d3df06060
Closes-Bug: #1454880