Updated openstack/openstack Project: openstack/governance  919984895c46f1046dde77051c6475267ca5e410
Adding the Chef cookbooks to OpenStack

This in the commit to added the Chef cookbooks to the OpenStack name
space.

Our meetings and office hours are twice weekly[1], on IRC, and our
roadmap is as transparent and managed on Launchpad again using standard
OpenStack community practices[3].

As of 2015-04-27 our community has decided to depreciate the Monday
Hangout in favor for IRC. The youtube channel[2] will be there for
historical tracking, but we will leverage the IRC logging here[4].

This patch is an official request to be part of the big tent and become
an official project.

[1] https://wiki.openstack.org/wiki/Chef/GettingStarted#Communication
[2] https://www.youtube.com/channel/UCPQhSl-wxgWJH6_r7pk5PbQ/videos
[3] https://wiki.openstack.org/wiki/Chef
[4] http://eavesdrop.openstack.org/meetings/openstack_chef/

Change-Id: I18f507da315951352e31a9d1be2aeb0e40c9d085