Updated openstack/openstack Project: openstack-infra/shade  39eefc1989ff4e51db8716dd7106967b14fee12a
Make sure glance image list actually runs in Tasks

images.list() does not actually talk to an API. So putting it in
TaskManager actually takes an execution slot that it does not need. On
the other hand, the follow up list expansion DOES talk to the API. So
put it in the Task, since it's the evil thing.

Change-Id: I7f30ba908552fbecaedd364ecf20b19e096b03d9