Updated openstack/openstack Project: openstack-infra/shade  ac94e8d9e193df569c93d0ec919d1f039f59bd28
Don't cache keystone tokens as KSC does it for us

We don't need to keep track of what our current valid keystone token is
becauase keystone client does it for us. Instead defer to
ksc.session.get_token() to either reuse a valid token or make a new one
for us if necessary.

Change-Id: I23f90b6292f32ce9cef962bdc641c291e4d2fe4d