Updated openstack/openstack Project: openstack/training-guides  21155d91de514b6379982005e856c1a6b9e90c44
labs: repeat test from snapshot

With this changeset, the cluster build can be tested starting from a
specific snapshot. This is mostly useful for testing client scripts
without having to go through the whole cluster build every time.

For instance, to rebuild the cluster from snapshot neuntron_configured:

./tools/repeat-test.sh -b -t neutron_configured -s "controller compute"

This will restore the appropriate snapshot on each node, boot nodes
controller and compute, and continue the cluster build from the named
snapshot onward.

Note: this procedure restores, boots, and continues in short order which
uncovers new shortcomings (races) in scripts that assume a service is
available when they should check and wait for the service.

If you just want to restore all nodes to the appropriate snapshots:
./tools/restore-cluster.sh -t <snapshot>

Change-Id: I0b7888d02de6eb50760d421d6d897f516f8c7718