Updated openstack/openstack Project: openstack/trove  7a68ec634b4ad2331ead123cbe330a366aa2eca5
accepting network and availability zone for instances in cluster

Standalone instance create call in trove already accepts neutron networks
and availability zone information.

In the similar manner, instances which are part of the cluster can
now be supplied with a specific neutron network information and
availability zone.

This review is addressing the fix only for vertica datastore.

Partial-Bug: #1447350
Change-Id: I5a1e9300dd0a16395769acda672943899f9ad0ad