Updated openstack/openstack Project: openstack/tripleo-incubator  0ff2dd2babda25ec79eda0bf4769c22543d3473a
Rename brbm.xml template to net.xml

The brbm.xml template is not used for multiple networks.
Additionally libvirt calls these net (or networks) so
renaming the template makes it less confusing.

Using a more explicit %NETWORK_NAME% tag in the template so
that it is more clear what is being replaced by a variable.

Change-Id: I066a5d0022871804479c1ee325881381b08020c6