Updated openstack/openstack Project: openstack-infra/irc-meetings  422b18b338a80a18cb6467357e3f0248fe4b1b3d
Remove Networking Guide Update Meeting

This meeting is in a Google Hangout, but this repository lists
IRC meetings.  The OpenStack project asks constituent projects
to use IRC as a primary forum for regular meetings so that they
are accessible to the whole community, participants can easily
join meetings casually as needed, and logs of meetings are
available for people unable to attend.

See http://governance.openstack.org/reference/irc.html for more
detail.

Change-Id: Ia39eaac3481919f968fbafdbcef21f62a6885e31