Updated openstack/openstack Project: openstack/openstack-doc-tools  8a8ccdfb8e8c94af9ed7b63cf392d66f7783c952
let the autohelp_wrapper use local repos

The autohelp_wrapper uses as git URL for the projects the URL
git://git.openstack.org/openstack. This is not configurable.
Because of that, the build of guides (e.g. the configuration
reference) based on local (not yet commited) branches was not
possible.
This change set offers a new parameter for the script to use
another URL for the projects (e.g. to a local one).

Example:
The command to build the config options table of a local nova branch
(with the name "docs_serial_console") which is in the local folder
"~/git" would look like this:
    ./autohelp-wrapper -g ~/git -b docs_serial_console docbook nova

Change-Id: I03e5491221ac03be596b243183bdeb98c63dd03e