Updated openstack/openstack Project: openstack/ironic  ed3030a49a5c75ec35f6f6856d5c2f6cdedd7014
Fix: iSCSI iqn name RFC violation

Pior to this patch the name generated by Ironic for the iSCSI target was
violating the RFC 3720 and RFC 3721, and this was messing up with some
tools consuming it (see deploy-targetcli element for diskimage-builder).

This patch fixing this generating a name that complies to the RFCs.

Change-Id: I27ae1321b4195483ad50c0a46e6912152a25af76
Closes-Bug: #1459243