Updated openstack/openstack Project: openstack/tripleo-heat-templates  a58431e4c7a94ae9fa9320e628fbf8aff882b183
Map Swift services to isolated networks

This change adds paramters to specify which networks the Swift API
services will use. If the storage network exists, it will be used
for the Swift API, otherwise the Undercloud 'ctlplane' network will
be used. If the storage_mgmt network exists, it will be used for
the back-end storage services, otherwise the 'ctlplane' will be
used by default.

Change-Id: I1d5e966a16416c52935c22efe2d4783cd2192c32