Updated openstack/openstack Project: openstack/tempest  7068d88a65a374e82cf982bd7add4044f3567e40
Fix sample conf file based on new oslo.log release

This commit updates the tempest sample config file to
include changes in the latest oslo.log release, 1.2.0.
This is needed to unblock the sample config generation
job which is failing because the in-tree tempest sample
config file differs from what is generated with a fresh
venv.

Change-Id: Id40310712ecda13d65cfd68bdaf4ea91dc7687fe