Updated openstack/openstack Project: openstack/keystone  26ebd4f284f43e1512a21eb4cfc006a74098184a
Move endpoint policy into keystone core

Remove endpoint_policy as an extension and move it to a core resource.
For now we leave the database migrations in the extension directory
until we have a general policy for merging these into core.

DocImpact: You should no longer run the migrations for this extension.
Implements: bp replace-extensions
Change-Id: I6215b7df497c142a5e73b62543e0d76458c85f64