Updated openstack/openstack Project: openstack/neutron  726a59c7d29f11df4cad0830b8dd9e8ba830ffa9
Flesh out the new RFE process and set deadlines for it's use

The new RFE process is great in concept, but as was discovered in the
first neutron-drivers meeting where we discussed these, there exist
some rough edges. Specifically around deadlines and the conversion to
using RFEs, the gray area was very obvious. This patch attempts to put
a stake in the ground for when we transition fully to this new model,
including distinct timelines.

Given that we will need to work with people during the transition,
what is proposed is a way to let us do that while not blocking existing
specs and work.

Change-Id: Ife50f65caf4b13405ad60dd4be3d347aa31ef3a7