Updated openstack/openstack Project: openstack/python-keystoneclient  795b8567174f1d210eab6a4585d7339a302b0a69
Fixed grammatical errors in the V2 Client API doc

In the using-api-v2.rst document, various inconsistencies in spelling of
"openstackDemo", as well as the password under the "Creating Tenants" and
"Creating Users" sections.

Updated service names to adhere to the OpenStack service naming conventions.

Fixed capitalization of "Client" to "client".

Change-Id: Ib20782f5b05f7097158f606c6562f92126650b89
Closes-Bug: #1458015