Updated openstack/openstack Project: openstack/oslo.messaging  baddce34a2f94487f162014df7423a63d58fcd18
Adding Publisher Acknowledgements/confirms

During the vancouver summit, Michael Klishin from
pivotal proposed we should use publisher confirms
as the default setting to reduce messages being
dropped on the floor.

https://www.rabbitmq.com/confirms.html
http://www.rabbitmq.com/blog/2011/02/10/introducing-publisher-confirms/

This setting is a lightweight way of keeping track of which messages
have been processed by the broker and which would need re-publishing
in case of broker shutdown or network failure.

Change-Id: I25bc955df130dad4725f5281211d37fd73e7ea63