Updated openstack/openstack Project: openstack-infra/irc-meetings  cd2c1f25e0f64126d48ededf6f51e762078a8034
Rename Gantt meeting to Nova Scheduler

As discussed in [1] we no longer want to use the word 'Gantt'
for describing our efforts related to the scheduler (including
the split).

That's only when the real fork will happen back in the Gantt
tree that we will again use the project name.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-March/060179.html

Change-Id: I972ae46efe60ce47ec9c63a8412696d2f2724b86