Updated openstack/openstack Project: openstack/keystone  5b650ff3f9b86c3aa6889fca479be8be57fc08e0
Don't assume group IDs are UUID format

When using the Fernet token provider to build a federated payload, we can't
assume that the group ID will always be a UUID format that we can convert to
bytes. This change makes the Fernet code smart enough to pass the original
group ID in the payload if it can't convert it to bytes.

Change-Id: I6d00902eb461c22aafd9cb5ca706b05bedefd37d
Related-Bug: 1459382