Updated openstack/openstack Project: openstack-infra/project-config  7017374d0d9b1b4e53dc90e8110c564c99f7a7cb
Don't include system test gems when doing lint

The system tests gems (which right now are
beaker, beaker-rspec, & serverspec) include many dependencies that add
weight and time to the bundle installation. If any system test gems are
not needed for the lint builder job, then they can be added to the
bundler group "system_tests" and will not be installed as part of the
lint run.

This is also how Puppet Labs configures their modules' Gemfiles.

Change-Id: I19239d6ca451515bd0d51f167e5984fd8f31d793