Updated openstack/openstack Project: openstack/ha-guide  039fb10ad54b831616c38ea9d70265c30670cc3b
How to configure haproxy

This is a slightly edited version of the information in the old ha-guide,
based on the material in https://review.openstack.org/#/c/164998/.

This section needs more information to explain the sample file that is provided
and instructions for restarting HAProxy after modifying the file.

Also we wanted a mention of commercial load-balancers added to this section;
please provide appropriate wording.

Change-Id: Ie5b346e496dc987eb9f228dc965008bc73d822d1