Updated openstack/openstack Project: openstack/neutron  9b0577ed4379ad78f4fe6f4f01f79aa76dc9debe
Isolate use of fixed_ips[0] to avoid confusion

I was looking at the assumption around using fixed_ips[0] in a recent
patch [1].  I thought at the least, the usage of fixed_ips[0] should
be isolated with the comment explaining why it is okay.

I thought that we could also use this patch as an opportunity to vet
the validity of the claim made in the comment.

[1] If4a310da06f9b0076a9f62926a16b574a8c109ce

Change-Id: Iba5713dd7d852429997ff43d98266a9f022d5d86