Updated openstack/openstack Project: openstack/ironic  d1b586cbf673af3c05257bd16339aadc0f61d7f4
Get admin auth token for Glance client in image_service

In some cases we do not have auth_token in context, but Glance client
requires it for authentication. This patch moves usage of Keystone
to image_service and adds check for Glance auth strategy. This
changes fix problems with Ironic standalone mode and reduces quantity
of Keystone API calls.

Change-Id: I5c625db6f30a76da10d72c4d429d3f3cbfd6f432