Updated openstack/openstack Project: openstack/trove  0f3d24254811906810552217cf05ed9409f415fc
Move guestagent settings to default section

The guestagent related settings are in the database section of the
taskmanager.conf.sample file.  If a user were to use this sample then
the guestagent related settings will not be read by the taskmanager
service because it expects these guestagent related settings to be
in the default section.  This fix moves the guestagent related
settings to the default section.

Change-Id: I8dbdf7abcfd5f2341831cc6daaf5cb14d8d42d61
Closes-Bug: #1459415